Change format methods into normal methods
Previously, it was necessary to explicitly pass the instance of the
calling item into the format method. This is because the format method
was passed in by the parent item, so a normal method would have the
instance of the parent item passed in. Since this is no longer the case,
the format methods have been made into normal methods.